chat/api/requirements.txt
fastapi
uvicorn
openai
chromadb
python-dotenv
pandas
tiktoken